Move proto conversion methods beside their types
Instead of having a separate proto.go file, place each Proto() method
and FromProto function directly in the file containing its type:
- task.go: Instruction, Task, RunnerEvent proto conversions
- link.go: Link proto conversions
- event.go: Event proto conversions
- log.go: Log proto conversion
