+++ internal/model/log.go
@@ -1,6 +1,10 @@
 package model
 
-import "time"
+import (
+	"time"
+
+	xagentv1 "github.com/icholy/xagent/internal/proto/xagent/v1"
+)
 
 // Log represents a log entry for a task.
 type Log struct {
@@ -9,4 +13,12 @@ type Log struct {
 	Type      string    `json:"type"`
 	Content   string    `json:"content"`
 	CreatedAt time.Time `json:"created_at"`
+}
+
+// Proto converts a Log to its protobuf LogEntry representation.
+func (l *Log) Proto() *xagentv1.LogEntry {
+	return &xagentv1.LogEntry{
+		Type:    l.Type,
+		Content: l.Content,
+	}
 }
